requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
astroid==3.2.2
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dill==0.3.8
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
identify==2.5.36
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.35.10
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pdfminer.six==20240706
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==5.27.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pylint==3.2.5
PyPDF2==3.0.1
pytest==8.2.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.36.0
sympy==1.12.1
tenacity==8.4.2
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
XlsxWriter==3.2.0
